Fixed migrations

+++ database/migrations/2023_10_18_160122_create_blog_category_translations_table.php
@@ -12,19 +12,21 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create('blog_category_translations', function (Blueprint $table) {
+        if (!Schema::hasTable('blog_category_translations')) {
+            Schema::create('blog_category_translations', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('blog_category_id');
             $table->string('locale', 191)->index();
             $table->string('title');
 
             $table->foreign('blog_category_id')->references('id')->on('blog_categories')->onDelete('cascade');
-        });
+            });
+        }
 
         $this->moveTitles();
 
         Schema::table('blog_categories', function (Blueprint $table) {
-            $table->dropColumn('title');
+            //$table->dropColumn('title');
         });
     }
